Shikadi now can toast you! he is in!

+++ src/engine/galaxy/ep5/ai/CShikadi.cpp
@@ -8,7 +8,8 @@
 
 #include "CShikadi.h"
 #include "engine/galaxy/common/ai/CPlayerBase.h"
-#include <engine/galaxy/common/ai/CPlayerLevel.h>
+#include "engine/galaxy/common/ai/CPlayerLevel.h"
+#include <engine/galaxy/common/ai/CEnemyShot.h>
 #include "misc.h"
 
 /*
@@ -23,7 +24,7 @@ $2DE2W #Shikadi walk
 $2E00W #Shikadi polezaps 8
 $2E1EW #Shikadi polezaps
 $2E3CW #Shikadi stunned 10
-$2E5AW #Shikadi polezap 11 -> TODO: I think this must be another type of object
+$2E5AW #Shikadi polezap 11
 $2E78W #Shikadi polezap
  */
 
@@ -34,30 +35,42 @@ enum SHIKADIACTIONS
 {
 A_SHIKADI_STAND = 0,
 A_SHIKADI_WALK = 4,
-A_SHIKADI_TURN = 4,
 A_SHIKADI_POLE_ZAPS = 8,
 A_SHIKADI_STUNNED = 10
 };
 
 const int WALK_SPEED = 25;
 
+const int TIME_UNTIL_STAND = 150;
+
+const int TIME_TO_ZAP = 30;
   
 CShikadi::CShikadi(CMap *pmap, const Uint16 foeID, const Uint32 x, const Uint32 y) :
 CStunnable(pmap, foeID, x, y),
+mHealth(4),
 mTimer(0)
 {
-	//mActionMap[A_AMPTON_STUNNED] = &CStunnable::processGettingStunned;
+    mActionMap[A_SHIKADI_STAND] = (void (CStunnable::*)()) &CShikadi::processStanding;
+    mActionMap[A_SHIKADI_WALK] = (void (CStunnable::*)()) &CShikadi::processWalking;
+    mActionMap[A_SHIKADI_POLE_ZAPS] = (void (CStunnable::*)()) &CShikadi::processPoleZaps;
+    mActionMap[A_SHIKADI_STUNNED] = &CStunnable::processGettingStunned;
   
-	// Adapt this AI
-	setupGalaxyObjectOnMap(0x2D10, A_SHIKADI_STAND);
-	
-	xDirection = LEFT;
+    // Adapt this AI
+    setupGalaxyObjectOnMap(0x2D10, A_SHIKADI_STAND);
+    
+    xDirection = LEFT;
 }
 
-
+void CShikadi::processStanding()
+{
+    if(getActionStatus(A_SHIKADI_WALK))
+    {
+	setAction(A_SHIKADI_WALK);
+    }
+}
 
 void CShikadi::processWalking()
-{
+{    
   // Move normally in the direction
   if( xDirection == RIGHT )
   {
@@ -67,6 +80,52 @@ void CShikadi::processWalking()
   {
     moveLeft( WALK_SPEED );
   }   
+  
+  if( blockedl )
+  {
+      xDirection = RIGHT;
+  }
+  else if(blockedr)
+  {
+      xDirection = LEFT;
+  }
+  
+  mTimer++;
+  if( mTimer < TIME_UNTIL_STAND )  
+      return;
+  
+  mTimer = 0;
+  
+  int newX = (xDirection == LEFT) ? getXLeftPos()+(4<<STC) : getXRightPos()-(4<<STC);
+  
+  if(hitdetectWithTilePropertyVert(1, newX, getYUpPos(), getYDownPos(), (1<<STC) ))
+  {
+    if(getProbability(700))
+    {	
+	newX = (newX>>CSF)<<CSF;
+	// Spawn a Enemyshot in form electrostatic  charge which goes upwards
+	g_pBehaviorEngine->m_EventList.spawnObj( new CEnemyShot(mp_Map, 0, 
+								newX, getYUpPos(),
+								0x2E5A, 0, UP,  150) );
+	setAction(A_SHIKADI_POLE_ZAPS); 
+	playSound(SOUND_POLEZAP);
+	return;
+    }      
+  }
+
+  if(getProbability(300))
+  {
+    setAction(A_SHIKADI_STAND);
+  }
+
+}
+
+void CShikadi::processPoleZaps()
+{    
+    if(getActionStatus(A_SHIKADI_WALK))
+    {
+	setAction(A_SHIKADI_WALK);
+    }
 }
 
 
@@ -96,15 +155,21 @@ void CShikadi::getTouchedBy(CSpriteObject &theObject)
 	// Was it a bullet? Than make it stunned.
 	if( dynamic_cast<CBullet*>(&theObject) )
 	{
-		playSound(SOUND_ROBO_STUN);
+	    mHealth--;
+	    theObject.dead = true;
+	    
+	    // TODO: Flash effect must be added here!
+	    if(mHealth == 0)
+	    {
+		setAction(A_SHIKADI_STUNNED);
 		dead = true;
-		theObject.dead = true;
+	    }
 	}
 
-	/*if( CPlayerBase *player = dynamic_cast<CPlayerBase*>(&theObject) )
+	if( CPlayerBase *player = dynamic_cast<CPlayerBase*>(&theObject) )
 	{
 		player->kill();
-	}*/
+	}
 }
 
 
@@ -121,15 +186,6 @@ void CShikadi::process()
 	performCollisions();
 	
 	performGravityMid();
-
-	if( blockedl )
-	{
-	  xDirection = RIGHT;
-	}
-	else if(blockedr)
-	{
-	  xDirection = LEFT;
-	}
 
 	if(!processActionRoutine())
 	    exists = false;
